backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
numpy>=1.26.0
scikit-learn>=1.5.0
umap-learn>=0.5.6
radon>=6.0.1
sentence-transformers>=3.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
